Split Q1Veritas bot into persona bot and convert-to-binary bot. Changed the perplexity model. Fixed scratch pad bug
