+++ .env.template
@@ -1,14 +1,18 @@
 PYTHONPATH=.
 
-# Currently not being used as models, but might be in the future
-PERPLEXITY_API_KEY=
+
+# As of Jan 23rd 2025, OpenAI and Exa are the most heavily used. Most bots only use a subset of these services
 OPENAI_API_KEY=
 EXA_API_KEY=
+PERPLEXITY_API_KEY=
+DEEPSEEK_API_KEY=
+ASKNEWS_CLIENT_ID=
+ASKNEWS_SECRET=
 
 # Fill this in if using the Metaculus API
 METACULUS_TOKEN=
 
-# Right now only used for free semantic similarity calculation in Deduplicator, but defaults to OpenAI if not filled in
+# AS of Jan 23rd 2025, only used for free semantic similarity calculation in Deduplicator, but defaults to OpenAI if not filled in
 HUGGINGFACE_API_KEY=
 
 # Disable if in Streamlit Cloud
