+++ forecasting_tools/ai_models/gpt4o.py
@@ -1,16 +1,16 @@
 import logging
 from typing import Final
 
-from forecasting_tools.ai_models.model_archetypes.openai_text_model import (
-    OpenAiTextToTextModel,
+from forecasting_tools.ai_models.model_archetypes.general_llm import (
+    GeneralTextToTextLlm,
 )
 
 logger = logging.getLogger(__name__)
 
 
-class Gpt4o(OpenAiTextToTextModel):
+class Gpt4o(GeneralTextToTextLlm):
     # See OpenAI Limit on the account dashboard for most up-to-date limit
-    MODEL_NAME: Final[str] = "gpt-4o"
+    MODEL_NAME: Final[str] = "openai/gpt-4o"
     REQUESTS_PER_PERIOD_LIMIT: Final[int] = 8_000
     REQUEST_PERIOD_IN_SECONDS: Final[int] = 60
     TIMEOUT_TIME: Final[int] = 40
