+++ forecasting_tools/ai_models/perplexity.py
@@ -1,12 +1,12 @@
 from typing import Final
 
-from forecasting_tools.ai_models.model_archetypes.perplexity_text_model import (
-    PerplexityTextModel,
+from forecasting_tools.ai_models.model_archetypes.general_llm import (
+    GeneralTextToTextLlm,
 )
 
 
-class Perplexity(PerplexityTextModel):
-    MODEL_NAME: Final[str] = "llama-3.1-sonar-huge-128k-online"
+class Perplexity(GeneralTextToTextLlm):
+    MODEL_NAME: Final[str] = "perplexity/sonar-pro"
     REQUESTS_PER_PERIOD_LIMIT: Final[int] = (
         40  # Technically 50, but giving wiggle room
     )
@@ -14,5 +14,3 @@ class Perplexity(PerplexityTextModel):
     TIMEOUT_TIME: Final[int] = 120
     TOKENS_PER_PERIOD_LIMIT: Final[int] = 2000000
     TOKEN_PERIOD_IN_SECONDS: Final[int] = 60
-    PRICE_PER_TOKEN: Final[float] = 0.000005
-    PRICE_PER_REQUEST: Final[float] = 0.005
