+++ scripts/benchmark_forecast_bot.py
@@ -8,17 +8,11 @@ import typeguard
 from forecasting_tools.ai_models.resource_managers.monetary_cost_manager import (
     MonetaryCostManager,
 )
-from forecasting_tools.forecasting.forecast_bots.experiments.q4v_w_exa_and_dseekr1 import (
-    Q4VeritasWithExaAndDeepSeekR1,
-)
 from forecasting_tools.forecasting.forecast_bots.forecast_bot import (
     ForecastBot,
 )
 from forecasting_tools.forecasting.forecast_bots.official_bots.q1_veritas_bot import (
     Q1VeritasBot,
-)
-from forecasting_tools.forecasting.forecast_bots.official_bots.q3_template_bot import (
-    Q3TemplateBot,
 )
 from forecasting_tools.forecasting.helpers.benchmarker import Benchmarker
 from forecasting_tools.util.custom_logger import CustomLogger
@@ -27,22 +21,22 @@ logger = logging.getLogger(__name__)
 
 
 async def benchmark_forecast_bot() -> None:
-    questions_to_use = 65
+    questions_to_use = 1
     with MonetaryCostManager() as cost_manager:
         bots = [
-            Q3TemplateBot(),
-            Q4VeritasWithExaAndDeepSeekR1(
-                research_reports_per_question=1,
-                predictions_per_research_report=1,
-            ),
-            Q4VeritasWithExaAndDeepSeekR1(
-                research_reports_per_question=5,
-                predictions_per_research_report=5,
-            ),
-            Q1VeritasBot(
-                research_reports_per_question=1,
-                predictions_per_research_report=1,
-            ),
+            # Q3TemplateBot(),
+            # Q4VeritasWithExaAndDeepSeekR1(
+            #     research_reports_per_question=1,
+            #     predictions_per_research_report=1,
+            # ),
+            # Q4VeritasWithExaAndDeepSeekR1(
+            #     research_reports_per_question=5,
+            #     predictions_per_research_report=5,
+            # ),
+            # Q1VeritasBot(
+            #     research_reports_per_question=1,
+            #     predictions_per_research_report=1,
+            # ),
             Q1VeritasBot(
                 research_reports_per_question=5,
                 predictions_per_research_report=5,
